refactor: 로또 게임 수행 메서드들 game.js 로  분리
